+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to the program will be documented here.
 - Quick-fixed error on startup when logs dir does not exist, long term solution required.
 - Updated README.md to reflect last update date.
 - Fixed changelog entry to show "unreleased" instead of 0.2.2.
+- Quick fix on monitor ignore/clear system.
 
 #### Changed